ct_cover (common_test v1.27)

Common Test framework code coverage support module.

Common Test framework code coverage support module.

This module exports help functions for performing code coverage analysis.

Functions

add_nodes/1

add_nodes(Nodes) -> {ok, StartedNodes} | {error, Reason}

Adds nodes to current cover test. Notice that this only works if cover support is active.

To have effect, this function is to be called from init_per_suite/1 (see ct_suite) before any tests are performed.

cross_cover_analyse(Level, Tests)
(since OTP R16B)

cross_cover_analyse(Level, Tests) -> ok

Accumulates cover results over multiple tests. See section Cross Cover Analysis in the Users's Guide.

remove_nodes/1

remove_nodes(Nodes) -> ok | {error, Reason}

Removes nodes from the current cover test.

Call this function to stop cover test on nodes previously added with ct_cover:add_nodes/1. Results on the remote node are transferred to the Common Test node.