ct_cover (common_test v1.20.5)

Common Test framework code coverage support module.

This module exports help functions for performing code coverage analysis.

Functions

add_nodes(Nodes) -> {ok, StartedNodes} | {error, Reason}

Nodes = [atom()]

StartedNodes = [atom()]

Reason = cover_not_running | not_main_node

Adds nodes to current cover test. Notice that this only works if cover support is active.

To have effect, this function is to be called from init_per_suite/1 (see ct_suite) before any tests are performed.

cross_cover_analyse(Level, Tests) -> ok
(since OTP R16B)

Level = overview | details

Tests = [{Tag, Dir}]

Tag = atom()

Dir = string()

Accumulates cover results over multiple tests. See section Cross Cover Analysis in the Users's Guide.

remove_nodes(Nodes) -> ok | {error, Reason}

Nodes = [atom()]

Reason = cover_not_running | not_main_node

Removes nodes from the current cover test.

Call this function to stop cover test on nodes previously added with ct_cover:add_nodes/1. Results on the remote node are transferred to the Common Test node.