ct_master (common_test v1.27)

Distributed test execution control for Common Test.

Distributed test execution control for Common Test.

This module exports functions for running Common Test nodes on multiple hosts in parallel.

Functions

abort()

abort() -> ok

Stops all running tests.

abort/1

abort(Nodes) -> ok

Stops tests on specified nodes.

basic_html(Bool)
(since OTP R15B01)

basic_html(Bool) -> ok

If set to true, the ct_master logs are written on a primitive HTML format, not using the Common Test CSS style sheet.

get_event_mgr_ref()
(since OTP 17.5)

get_event_mgr_ref() -> MasterEvMgrRef

Gets a reference to the Common Test master event manager. The reference can be used to, for example, add a user-specific event handler while tests are running.

Example:

 gen_event:add_handler(ct_master:get_event_mgr_ref(), my_ev_h, [])

progress()

progress() -> [{Node, Status}]

Returns test progress. If Status is ongoing, tests are running on the node and are not yet finished.

run/1

run(TestSpecs) -> ok

Equivalent to ct_master:run(TestSpecs, false, [], []).

run(TestSpecs, InclNodes, ExclNodes)

run(TestSpecs, InclNodes, ExclNodes) -> ok

Equivalent to ct_master:run(TestSpecs, false, InclNodes, ExclNodes).

run/4

run(TestSpecs, AllowUserTerms, InclNodes, ExclNodes) -> ok

Tests are spawned on the nodes as specified in TestSpecs. Each specification in TestSpec is handled separately. However, it is also possible to specify a list of specifications to be merged into one specification before the tests are executed. Any test without a particular node specification is also executed on the nodes in InclNodes. Nodes in the ExclNodes list are excluded from the test.

run_on_node(TestSpecs, Node)

run_on_node(TestSpecs, Node) -> ok

Equivalent to ct_master:run_on_node(TestSpecs, false, Node).

run_on_node/3

run_on_node(TestSpecs, AllowUserTerms, Node) -> ok

Tests are spawned on Node according to TestSpecs.

run_test(Node, Opts)

run_test(Node, Opts) -> ok

Tests are spawned on Node using ct:run_test/1