ct_ssh (common_test v1.27)

SSH/SFTP client module.

SSH/SFTP client module.

This module uses application SSH, which provides detailed information about, for example, functions, types, and options.

Argument Server in the SFTP functions is only to be used for SFTP sessions that have been started on existing SSH connections (that is, when the original connection type is ssh). Whenever the connection type is sftp, use the SSH connection reference only.

The following options are valid for specifying an SSH/SFTP connection (that is, can be used as configuration elements):

 [{ConnType, Addr},
  {port, Port},
  {user, UserName}
  {password, Pwd}
  {user_dir, String}
  {public_key_alg, PubKeyAlg}
  {connect_timeout, Timeout}
  {key_cb, KeyCallbackMod}]

ConnType = ssh | sftp.

For other types, see ssh.

All time-out parameters in ct_ssh functions are values in milliseconds.

Types

connection()
(not exported)

-type connection() :: handle() | ct:target_name().

For target_name, see module ct.

handle()
(not exported)

-type handle() :: pid().

Handle for a specific SSH/SFTP connection, see module ct.

ssh_sftp_return()
(not exported)

-type ssh_sftp_return() :: term().

Return value from an ssh_sftp function.

Functions

apread(SSH, Handle, Position, Length)

-spec apread(SSH, Handle, Position, Length) -> Result
          when
              SSH :: connection(),
              Handle :: term(),
              Position :: integer(),
              Length :: integer(),
              Result :: ssh_sftp_return().

apread(SSH, Handle, Position, Length) -> Result

For information and other types, see ssh_sftp.

apread(SSH, Server, Handle, Position, Length)

apread(SSH, Server, Handle, Position, Length) -> Result

For information and other types, see ssh_sftp.

apwrite(SSH, Handle, Position, Data)

apwrite(SSH, Handle, Position, Data) -> Result

For information and other types, see ssh_sftp.

apwrite(SSH, Server, Handle, Position, Data)

apwrite(SSH, Server, Handle, Position, Data) -> Result

For information and other types, see ssh_sftp.

aread(SSH, Handle, Len)

aread(SSH, Handle, Len) -> Result

For information and other types, see ssh_sftp.

aread(SSH, Server, Handle, Len)

aread(SSH, Server, Handle, Len) -> Result

For information and other types, see ssh_sftp.

awrite(SSH, Handle, Data)

awrite(SSH, Handle, Data) -> Result

For information and other types, see ssh_sftp.

awrite(SSH, Server, Handle, Data)

awrite(SSH, Server, Handle, Data) -> Result

For information and other types, see ssh_sftp.

close(SSH, Handle)

close(SSH, Handle) -> Result

For information and other types, see ssh_sftp.

close(SSH, Server, Handle)

close(SSH, Server, Handle) -> Result

For information and other types, see ssh_sftp.

connect(KeyOrName)

connect(KeyOrName) -> {ok, Handle} | {error, Reason}

Equivalent to ct_ssh:connect(KeyOrName, host, []).

connect/2

connect(KeyOrName, ConnType) -> {ok, Handle} | {error, Reason}

Equivalent to ct_ssh:connect(KeyOrName, ConnType, []).

connect(KeyOrName, ConnType, ExtraOpts)

connect(KeyOrName, ConnType, ExtraOpts) -> {ok, Handle} | {error, Reason}

Opens an SSH or SFTP connection using the information associated with KeyOrName.

If Name (an alias name for Key) is used to identify the connection, this name can be used as connection reference for subsequent calls. Only one open connection at a time associated with Name is possible. If Key is used, the returned handle must be used for subsequent calls (multiple connections can be opened using the configuration data specified by Key).

For information on how to create a new Name, see ct:require/2.

For target_name, see module ct.

ConnType always overrides the type specified in the address tuple in the configuration data (and in ExtraOpts). So it is possible to, for example, open an SFTP connection directly using data originally specifying an SSH connection. Value host means that the connection type specified by the host option (either in the configuration data or in ExtraOpts) is used.

ExtraOpts (optional) are extra SSH options to be added to the configuration data for KeyOrName. The extra options override any existing options with the same key in the configuration data. For details on valid SSH options, see application SSH.

del_dir(SSH, Name)

del_dir(SSH, Name) -> Result

For information and other types, see ssh_sftp.

del_dir(SSH, Server, Name)

del_dir(SSH, Server, Name) -> Result

For information and other types, see ssh_sftp.

delete(SSH, Name)

delete(SSH, Name) -> Result

For information and other types, see ssh_sftp.

delete(SSH, Server, Name)

delete(SSH, Server, Name) -> Result

For information and other types, see ssh_sftp.

disconnect(SSH)

disconnect(SSH) -> ok | {error, Reason}

Closes an SSH/SFTP connection.

exec(SSH, Command)

exec(SSH, Command) -> {ok, Data} | {error, Reason}

Equivalent to ct_ssh:exec(SSH, Command, DefaultTimeout).

exec/3

exec(SSH, Command, Timeout) -> {ok, Data} | {error, Reason}

Requests server to perform Command. A session channel is opened automatically for the request. Data is received from the server as a result of the command.

exec(SSH, ChannelId, Command, Timeout)

exec(SSH, ChannelId, Command, Timeout) -> {ok, Data} | {error, Reason}

Requests server to perform Command. A previously opened session channel is used for the request. Data is received from the server as a result of the command.

get_file_info(SSH, Handle)

get_file_info(SSH, Handle) -> Result

For information and other types, see ssh_sftp.

get_file_info(SSH, Server, Handle)

get_file_info(SSH, Server, Handle) -> Result

For information and other types, see ssh_sftp.

list_dir(SSH, Path)

list_dir(SSH, Path) -> Result

For information and other types, see ssh_sftp.

list_dir(SSH, Server, Path)

list_dir(SSH, Server, Path) -> Result

For information and other types, see ssh_sftp.

make_dir(SSH, Name)

make_dir(SSH, Name) -> Result

For information and other types, see ssh_sftp.

make_dir(SSH, Server, Name)

make_dir(SSH, Server, Name) -> Result

For information and other types, see ssh_sftp.

make_symlink(SSH, Name, Target)

make_symlink(SSH, Name, Target) -> Result

For information and other types, see ssh_sftp.

make_symlink(SSH, Server, Name, Target)

make_symlink(SSH, Server, Name, Target) -> Result

For information and other types, see ssh_sftp.

open(SSH, File, Mode)

open(SSH, File, Mode) -> Result

For information and other types, see ssh_sftp.

open(SSH, Server, File, Mode)

open(SSH, Server, File, Mode) -> Result

For information and other types, see ssh_sftp.

opendir(SSH, Path)

opendir(SSH, Path) -> Result

For information and other types, see ssh_sftp.

opendir(SSH, Server, Path)

opendir(SSH, Server, Path) -> Result

For information and other types, see ssh_sftp.

position(SSH, Handle, Location)

position(SSH, Handle, Location) -> Result

For information and other types, see ssh_sftp.

position(SSH, Server, Handle, Location)

position(SSH, Server, Handle, Location) -> Result

For information and other types, see ssh_sftp.

pread(SSH, Handle, Position, Length)

pread(SSH, Handle, Position, Length) -> Result

For information and other types, see ssh_sftp.

pread(SSH, Server, Handle, Position, Length)

pread(SSH, Server, Handle, Position, Length) -> Result

For information and other types, see ssh_sftp.

pwrite(SSH, Handle, Position, Data)

pwrite(SSH, Handle, Position, Data) -> Result

For information and other types, see ssh_sftp.

pwrite(SSH, Server, Handle, Position, Data)

pwrite(SSH, Server, Handle, Position, Data) -> Result

For information and other types, see ssh_sftp.

read(SSH, Handle, Len)

read(SSH, Handle, Len) -> Result

For information and other types, see ssh_sftp.

read(SSH, Server, Handle, Len)

read(SSH, Server, Handle, Len) -> Result

For information and other types, see ssh_sftp.

read_file(SSH, File)

read_file(SSH, File) -> Result

For information and other types, see ssh_sftp.

read_file(SSH, Server, File)

read_file(SSH, Server, File) -> Result

For information and other types, see ssh_sftp.

read_file_info(SSH, Name)

read_file_info(SSH, Name) -> Result

For information and other types, see ssh_sftp.

read_file_info(SSH, Server, Name)

read_file_info(SSH, Server, Name) -> Result

For information and other types, see ssh_sftp.

read_link(SSH, Name)

read_link(SSH, Name) -> Result

For information and other types, see ssh_sftp.

read_link(SSH, Server, Name)

read_link(SSH, Server, Name) -> Result

For information and other types, see ssh_sftp.

read_link_info(SSH, Name)

read_link_info(SSH, Name) -> Result

For information and other types, see ssh_sftp.

read_link_info(SSH, Server, Name)

read_link_info(SSH, Server, Name) -> Result

For information and other types, see ssh_sftp.

receive_response(SSH, ChannelId)

receive_response(SSH, ChannelId) -> {ok, Data} | {error, Reason}

Equivalent to ct_ssh:receive_response(SSH, ChannelId, close).

receive_response/3

receive_response(SSH, ChannelId, End) -> {ok, Data} | {error, Reason}

Equivalent to ct_ssh:receive_response(SSH, ChannelId, End, DefaultTimeout).

receive_response(SSH, ChannelId, End, Timeout)

receive_response(SSH, ChannelId, End, Timeout) -> {ok, Data} | {timeout, Data} | {error, Reason}

Receives expected data from server on the specified session channel.

If End == close, data is returned to the caller when the channel is closed by the server. If a time-out occurs before this happens, the function returns {timeout,Data} (where Data is the data received so far).

If End == timeout, a time-out is expected and {ok,Data} is returned both in the case of a time-out and when the channel is closed.

If End is a fun, this fun is called with one argument, the data value in a received ssh_cm message (see ssh_connection. The fun is to return either true to end the receiving operation (and have the so far collected data returned) or false to wait for more data from the server. Even if a fun is supplied, the function returns immediately if the server closes the channel).

rename(SSH, OldName, NewName)

rename(SSH, OldName, NewName) -> Result

For information and other types, see ssh_sftp.

rename(SSH, Server, OldName, NewName)

rename(SSH, Server, OldName, NewName) -> Result

For information and other types, see ssh_sftp.

send(SSH, ChannelId, Data)

send(SSH, ChannelId, Data) -> ok | {error, Reason}

Equivalent to ct_ssh:send(SSH, ChannelId, 0, Data, DefaultTimeout).

send/4

send(SSH, ChannelId, Data, Timeout) -> ok | {error, Reason}

Equivalent to ct_ssh:send(SSH, ChannelId, 0, Data, Timeout).

send(SSH, ChannelId, Type, Data, Timeout)

send(SSH, ChannelId, Type, Data, Timeout) -> ok | {error, Reason}

Sends data to server on specified session channel.

send_and_receive(SSH, ChannelId, Data)

send_and_receive(SSH, ChannelId, Data) -> {ok, Data} | {error, Reason}

Equivalent to ct_ssh:send_and_receive(SSH, ChannelId, Data, close).

send_and_receive/4

send_and_receive(SSH, ChannelId, Data, End) -> {ok, Data} | {error, Reason}

Equivalent to ct_ssh;send_and_receive(SSH, ChannelId, 0, Data, End, DefaultTimeout).

send_and_receive/5

send_and_receive(SSH, ChannelId, Data, End, Timeout) -> {ok, Data} | {error, Reason}

Equivalent to ct_ssh:send_and_receive(SSH, ChannelId, 0, Data, End, Timeout).

send_and_receive(SSH, ChannelId, Type, Data, End, Timeout)

send_and_receive(SSH, ChannelId, Type, Data, End, Timeout) -> {ok, Data} | {error, Reason}

Sends data to server on specified session channel and waits to receive the server response.

For details on argument End, see ct_ssh:receive_response/4.

session_close(SSH, ChannelId)

session_close(SSH, ChannelId) -> ok | {error, Reason}

Closes an SSH session channel.

session_open(SSH)

session_open(SSH) -> {ok, ChannelId} | {error, Reason}

Equivalent to ct_ssh:session_open(SSH, DefaultTimeout).

session_open(SSH, Timeout)

session_open(SSH, Timeout) -> {ok, ChannelId} | {error, Reason}

Opens a channel for an SSH session.

sftp_connect(SSH)

sftp_connect(SSH) -> {ok, Server} | {error, Reason}

Starts an SFTP session on an already existing SSH connection. Server identifies the new session and must be specified whenever SFTP requests are to be sent.

shell(SSH, ChannelId)
(since OTP 20.0)

-spec shell(SSH, ChannelId) -> Result
         when
             SSH :: handle() | ct:target_name(),
             ChannelId :: ssh:ssh_channel_id(),
             Result :: ok | {error, term()}.

shell(SSH, ChannelId) -> ok | {error, Reason}

Equivalent to ct_ssh:shell(SSH, ChannelId, DefaultTimeout).

shell(SSH, ChannelId, Timeout)
(since OTP 20.0)

-spec shell(SSH, ChannelId, Timeout) -> Result
         when
             SSH :: handle() | ct:target_name(),
             ChannelId :: ssh:ssh_channel_id(),
             Timeout :: timeout(),
             Result :: ok | {error, term()}.

shell(SSH, ChannelId, Timeout) -> ok | {error, Reason}

Requests that the user default shell (typically defined in /etc/passwd in Unix systems) is executed at the server end.

subsystem(SSH, ChannelId, Subsystem)

subsystem(SSH, ChannelId, Subsystem) -> Status | {error, Reason}

Equivalent to ct_ssh:subsystem(SSH, ChannelId, Subsystem, DefaultTimeout).

subsystem(SSH, ChannelId, Subsystem, Timeout)

subsystem(SSH, ChannelId, Subsystem, Timeout) -> Status | {error, Reason}

Sends a request to execute a predefined subsystem.

write(SSH, Handle, Data)

write(SSH, Handle, Data) -> Result

For information and other types, see ssh_sftp.

write(SSH, Server, Handle, Data)

write(SSH, Server, Handle, Data) -> Result

For information and other types, see ssh_sftp.

write_file(SSH, File, Iolist)

write_file(SSH, File, Iolist) -> Result

For information and other types, see ssh_sftp.

write_file(SSH, Server, File, Iolist)

write_file(SSH, Server, File, Iolist) -> Result

For information and other types, see ssh_sftp.

write_file_info(SSH, Name, Info)

write_file_info(SSH, Name, Info) -> Result

For information and other types, see ssh_sftp.

write_file_info(SSH, Server, Name, Info)

write_file_info(SSH, Server, Name, Info) -> Result

For information and other types, see ssh_sftp.