unix_telnet (common_test v1.27.3)

Callback module for ct_telnet, for connecting to a Telnet server on a UNIX host.

It requires the following entry in the configuration file:

{unix,[{telnet,HostNameOrIpAddress},
       {port,PortNum},                 % optional
       {username,UserName},
       {password,Password},
       {keep_alive,Bool}]}.            % optional

To communicate through Telnet to the host specified by HostNameOrIpAddress, use the interface functions in ct_telnet, for example, open(Name) and cmd(Name,Cmd).

Name is the name you allocated to the Unix host in your require statement, for example:

suite() -> [{require,Name,{unix,[telnet]}}].

or

ct:require(Name,{unix,[telnet]}).

The "keep alive" activity (that is, that Common Test sends NOP to the server every 10 seconds if the connection is idle) can be enabled or disabled for one particular connection as described here. It can be disabled for all connections using telnet_settings (see ct_telnet).

The {port,PortNum} tuple is optional and if omitted, default Telnet port 23 is used. Also the keep_alive tuple is optional, and the value default to true (enabled).

See Also

ct, ct_telnet

Functions

connect(ConnName, Ip, Port, Timeout, KeepAlive, TCPNoDelay, Extra)
(since OTP 18.3.3)

-spec connect(ConnName, Ip, Port, Timeout, KeepAlive, TCPNoDelay, Extra) ->
                 {ok, Handle} | {error, Reason}
                 when
                     ConnName :: ct:target_name(),
                     Ip :: inet:socket_address() | inet:hostname(),
                     Port :: inet:port_number(),
                     Timeout :: timeout(),
                     KeepAlive :: boolean(),
                     TCPNoDelay :: boolean(),
                     Extra :: {Username, Password} | KeyOrName,
                     Username :: iodata(),
                     Password :: iodata(),
                     KeyOrName :: ct:key_or_name(),
                     Handle :: ct_telnet:handle(),
                     Reason :: term().

Callback for ct_telnet.erl.

Setup Telnet connection to a Unix host.

get_prompt_regexp()

-spec get_prompt_regexp() -> Pattern when Pattern :: ct_telnet:prompt_regexp().

Callback for ct_telnet.erl.

Returns a suitable regexp string matching common prompts for users on Unix hosts.